Write traits to define protocols
