+++ server/src/protocol/mod.rs
@@ -2,5 +2,6 @@ pub mod deposit;
 pub mod ecdsa;
 pub mod ping;
 pub mod transfer;
+pub mod transfer_batch;
 pub mod util;
 pub mod withdraw;
